requirements.txt
fastapi
uvicorn
openai
langchain
langchain-community
python-dotenv
PyPDF2
beautifulsoup4
python-docx
tiktoken
